Rename ApiProvider to GraphQLApiProvider
